+++ src/variables.js
@@ -1,6 +1,0 @@
-var defaultUserLevel = 0; // this affects all users;
-exports.defaultUserLevel = defaultUserLevel;
-var vtcLevel = 0; // 0: regular / 1: premium / 3: special
-exports.vtcLevel = vtcLevel;
-var userLevel = -1;
-exports.userLevel = userLevel;