+++ changelog.md
@@ -67,6 +67,7 @@
 32. Replaced `vars.specialRoles` and `vars.specialRolesMap` with Context API `specialRoles` and `specialUsers`
 33. Replaced `vars.patrons` and `vars.userPatreonID` with Context API `patrons` and `curUserPatreonID`
 34. Replaced `vars.userConfig` with Context API `userConfig`
+35. Replaced `vars.vtc|userLevel` with Context API `vtc|userLevel`
 
 ## v3.3.x